requirements.txt
aiogram==3.15.0
paramiko==3.5.1
aiosqlite==0.20.0
python-dotenv==1.0.1
cryptography==44.0.0
